fix: #9580, proper 404 when ajaxifying
