+++ src/controllers/404.js
@@ -7,6 +7,7 @@ const validator = require('validator');
 const meta = require('../meta');
 const plugins = require('../plugins');
 const middleware = require('../middleware');
+const helpers = require('../middleware/helpers');
 
 exports.handle404 = function handle404(req, res) {
 	const relativePath = nconf.get('relative_path');
@@ -22,7 +23,13 @@ exports.handle404 = function handle404(req, res) {
 
 	if (isClientScript.test(req.url)) {
 		res.type('text/javascript').status(404).send('Not Found');
-	} else if (req.path.startsWith(`${relativePath}/assets/uploads`) || (req.get('accept') && !req.get('accept').includes('text/html')) || req.path === '/favicon.ico') {
+	} else if (
+		!res.locals.isAPI && (
+			req.path.startsWith(`${relativePath}/assets/uploads`) ||
+			(req.get('accept') && !req.get('accept').includes('text/html')) ||
+			req.path === '/favicon.ico'
+		)
+	) {
 		meta.errors.log404(req.path || '');
 		res.sendStatus(404);
 	} else if (req.accepts('html')) {
@@ -41,7 +48,11 @@ exports.send404 = async function (req, res) {
 	res.status(404);
 	const path = String(req.path || '');
 	if (res.locals.isAPI) {
-		return res.json({ path: validator.escape(path.replace(/^\/api/, '')), title: '[[global:404.title]]' });
+		return res.json({
+			path: validator.escape(path.replace(/^\/api/, '')),
+			title: '[[global:404.title]]',
+			bodyClass: helpers.buildBodyClass(req, res),
+		});
 	}
 
 	if (req.method === 'GET') {
@@ -49,5 +60,9 @@ exports.send404 = async function (req, res) {
 	}
 
 	await middleware.buildHeaderAsync(req, res);
-	await res.render('404', { path: validator.escape(path), title: '[[global:404.title]]' });
+	await res.render('404', {
+		path: validator.escape(path),
+		title: '[[global:404.title]]',
+		bodyClass: helpers.buildBodyClass(req, res),
+	});
 };
